add windows.core project refresh tile in task
I moved the windows logic to a new windows.core project to separate the
concerns of logic and view a bit better (even on the plattform level).
#363
