+++ Src/MoneyManager.Windows.Core/BackgroundTaskService.cs
@@ -5,7 +5,7 @@ using System.Threading.Tasks;
 using Windows.ApplicationModel.Background;
 using MoneyManager.Foundation.Interfaces;
 
-namespace MoneyManager.Windows
+namespace MoneyManager.Windows.Core
 {
     public class BackgroundTaskService : IBackgroundTaskService
     {
