+++ Src/MoneyManager.Windows.Core/MoneyManager.Windows.Core.csproj
@@ -4,11 +4,11 @@
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{F148873D-B85E-41C3-BE48-C69134F5F1CF}</ProjectGuid>
-    <OutputType>winmdobj</OutputType>
+    <ProjectGuid>{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}</ProjectGuid>
+    <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>MoneyManager.Tasks.Windows</RootNamespace>
-    <AssemblyName>MoneyManager.Tasks.Windows</AssemblyName>
+    <RootNamespace>MoneyManager.Windows.Core</RootNamespace>
+    <AssemblyName>MoneyManager.Windows.Core</AssemblyName>
     <DefaultLanguage>en-US</DefaultLanguage>
     <TargetPlatformIdentifier>UAP</TargetPlatformIdentifier>
     <TargetPlatformVersion>10.0.10240.0</TargetPlatformVersion>
@@ -16,7 +16,6 @@
     <MinimumVisualStudioVersion>14</MinimumVisualStudioVersion>
     <FileAlignment>512</FileAlignment>
     <ProjectTypeGuids>{A5A43C5B-DE2A-4C0C-9213-0A381AF9435A};{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}</ProjectTypeGuids>
-    <AllowCrossPlatformRetargeting>false</AllowCrossPlatformRetargeting>
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
     <PlatformTarget>AnyCPU</PlatformTarget>
@@ -108,31 +107,42 @@
     <None Include="project.json" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="ClearTransactionBackgroundTask.cs" />
+    <Compile Include="AppInformation.cs" />
+    <Compile Include="BackgroundTaskService.cs" />
+    <Compile Include="DatabasePath.cs" />
+    <Compile Include="DialogService.cs" />
+    <Compile Include="LocalSettings.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="RoamingSettings.cs" />
+    <Compile Include="Services\OneDriveBackupService.cs" />
+    <Compile Include="Shortcut\IncomeTile.cs" />
+    <Compile Include="Shortcut\SpendingTile.cs" />
+    <Compile Include="Shortcut\Tile.cs" />
+    <Compile Include="Shortcut\TransferTile.cs" />
+    <Compile Include="StoreFeatures.cs" />
+    <Compile Include="ThemedImageConverterLogic.cs" />
+    <Compile Include="TileHelper.cs" />
+    <Compile Include="UserNotification.cs" />
+    <Content Include="Properties\MoneyManager.Windows.Core.rd.xml" />
+  </ItemGroup>
+  <ItemGroup>
+    <SDKReference Include="LiveSDKXAML, Version=5.6">
+      <Name>Live SDK</Name>
+    </SDKReference>
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\MoneyManager.Core\MoneyManager.Core.csproj">
       <Project>{64f3a03f-44c2-44fa-bd29-fa23c35c95f0}</Project>
       <Name>MoneyManager.Core</Name>
     </ProjectReference>
-    <ProjectReference Include="..\MoneyManager.DataAccess\MoneyManager.DataAccess.csproj">
-      <Project>{038dfd1b-3b87-4098-8b16-8bc011b9eefc}</Project>
-      <Name>MoneyManager.DataAccess</Name>
-    </ProjectReference>
     <ProjectReference Include="..\MoneyManager.Foundation\MoneyManager.Foundation.csproj">
       <Project>{6936f4d3-8889-48c7-a676-80d59b2740ce}</Project>
       <Name>MoneyManager.Foundation</Name>
     </ProjectReference>
-    <ProjectReference Include="..\MoneyManager.Windows.Core\MoneyManager.Windows.Core.csproj">
-      <Project>{35e6de23-923d-40ce-b0cd-9ca27638fa5c}</Project>
-      <Name>MoneyManager.Windows.Core</Name>
+    <ProjectReference Include="..\MoneyManager.Localization\MoneyManager.Localization.csproj">
+      <Project>{90644219-756f-45cc-ab90-253463c5bc5e}</Project>
+      <Name>MoneyManager.Localization</Name>
     </ProjectReference>
-  </ItemGroup>
-  <ItemGroup>
-    <SDKReference Include="SQLite.UAP.2015, Version=3.9.2">
-      <Name>SQLite for Universal App Platform</Name>
-    </SDKReference>
   </ItemGroup>
   <PropertyGroup Condition=" '$(VisualStudioVersion)' == '' or '$(VisualStudioVersion)' &lt; '14.0' ">
     <VisualStudioVersion>14.0</VisualStudioVersion>
