+++ Src/MoneyManager.Windows.Core/Shortcut/IncomeTile.cs
@@ -4,11 +4,11 @@ using Windows.UI.StartScreen;
 using MoneyManager.Foundation.Interfaces.Shotcuts;
 using MoneyManager.Localization;
 
-namespace MoneyManager.Windows.Shortcut
+namespace MoneyManager.Windows.Core.Shortcut
 {
-    public class TransferTile : Tile, ITransferShortcut
+    public class IncomeTile : Tile, IIncomeShortcut
     {
-        public const string ID = "AddTransferTile";
+        public const string ID = "AddIncomeTile";
 
         public bool IsShortcutExisting => TileExists(ID);
 
@@ -16,9 +16,9 @@ namespace MoneyManager.Windows.Shortcut
         {
             await Create(new SecondaryTile(
                 ID,
-                Strings.AddTransferLabel,
+                Strings.AddIncomeLabel,
                 "intake",
-                new Uri("ms-appx:///Assets/TransferTileIcon.png", UriKind.Absolute),
+                new Uri("ms-appx:///Assets/IncomeTileIcon.png", UriKind.Absolute),
                 TileSize.Default));
         }
