+++ Src/MoneyManager.Windows.Core/StoreFeatures.cs
@@ -3,7 +3,7 @@ using Windows.ApplicationModel.Store;
 using Windows.System;
 using MoneyManager.Foundation.Interfaces;
 
-namespace MoneyManager.Windows
+namespace MoneyManager.Windows.Core
 {
     public class StoreFeatures : IStoreFeatures
     {
