+++ Src/MoneyManager.Windows.Core/ThemedImageConverterLogic.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using Windows.UI.Xaml;
 using Windows.UI.Xaml.Media.Imaging;
 
-namespace MoneyManager.Windows
+namespace MoneyManager.Windows.Core
 {
     public static class ThemedImageConverterLogic
     {
