+++ Src/MoneyManager.Windows.Tests/project.json
@@ -2,6 +2,7 @@
   "dependencies": {
     "Microsoft.NETCore.UniversalWindowsPlatform": "5.0.0",
     "MvvmCross.HotTuna.Plugin.SQLitePCL": "4.0.0-beta6",
+    "Xamarin.Insights": "1.11.2",
     "xunit": "2.1.0",
     "xunit.runner.visualstudio": "2.1.0"
   },
