+++ Src/MoneyManager.Windows.Tests/project.lock.json
@@ -16463,6 +16463,7 @@
     "": [
       "Microsoft.NETCore.UniversalWindowsPlatform >= 5.0.0",
       "MvvmCross.HotTuna.Plugin.SQLitePCL >= 4.0.0-beta6",
+      "Xamarin.Insights >= 1.11.2",
       "xunit >= 2.1.0",
       "xunit.runner.visualstudio >= 2.1.0"
     ],
