+++ Src/MoneyManager.Windows/App.xaml.cs
@@ -14,7 +14,8 @@ using Cirrious.MvvmCross.ViewModels;
 using MoneyManager.Core.Authentication;
 using MoneyManager.Core.Helpers;
 using MoneyManager.Core.ViewModels;
-using MoneyManager.Windows.Shortcut;
+using MoneyManager.Windows.Core;
+using MoneyManager.Windows.Core.Shortcut;
 using MoneyManager.Windows.Views;
 
 namespace MoneyManager.Windows
