+++ Src/MoneyManager.Windows/Setup.cs
@@ -7,8 +7,9 @@ using Cirrious.MvvmCross.ViewModels;
 using Cirrious.MvvmCross.WindowsUWP.Platform;
 using MoneyManager.Foundation.Interfaces;
 using MoneyManager.Foundation.Interfaces.Shotcuts;
-using MoneyManager.Windows.Services;
-using MoneyManager.Windows.Shortcut;
+using MoneyManager.Windows.Core;
+using MoneyManager.Windows.Core.Services;
+using MoneyManager.Windows.Core.Shortcut;
 using MvvmCross.Plugins.Email;
 using MvvmCross.Plugins.Email.WindowsCommon;
 using MvvmCross.Plugins.Sqlite;
