+++ Src/MoneyManager.sln
@@ -40,6 +40,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tasks", "Tasks", "{9705BAF6
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MoneyManager.Tasks.Windows", "MoneyManager.Tasks.Windows\MoneyManager.Tasks.Windows.csproj", "{F148873D-B85E-41C3-BE48-C69134F5F1CF}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MoneyManager.Windows.Core", "MoneyManager.Windows.Core\MoneyManager.Windows.Core.csproj", "{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Ad-Hoc|Any CPU = Ad-Hoc|Any CPU
@@ -704,6 +706,62 @@ Global
 		{F148873D-B85E-41C3-BE48-C69134F5F1CF}.Release|x64.Build.0 = Release|x64
 		{F148873D-B85E-41C3-BE48-C69134F5F1CF}.Release|x86.ActiveCfg = Release|x86
 		{F148873D-B85E-41C3-BE48-C69134F5F1CF}.Release|x86.Build.0 = Release|x86
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.Ad-Hoc|Any CPU.ActiveCfg = Release|Any CPU
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.Ad-Hoc|Any CPU.Build.0 = Release|Any CPU
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.Ad-Hoc|ARM.ActiveCfg = Release|ARM
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.Ad-Hoc|ARM.Build.0 = Release|ARM
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.Ad-Hoc|iPhone.ActiveCfg = Release|Any CPU
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.Ad-Hoc|iPhone.Build.0 = Release|Any CPU
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.Ad-Hoc|iPhoneSimulator.ActiveCfg = Release|Any CPU
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.Ad-Hoc|iPhoneSimulator.Build.0 = Release|Any CPU
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.Ad-Hoc|Mixed Platforms.ActiveCfg = Release|x86
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.Ad-Hoc|Mixed Platforms.Build.0 = Release|x86
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.Ad-Hoc|x64.ActiveCfg = Release|x64
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.Ad-Hoc|x64.Build.0 = Release|x64
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.Ad-Hoc|x86.ActiveCfg = Release|x86
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.Ad-Hoc|x86.Build.0 = Release|x86
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.AppStore|Any CPU.ActiveCfg = Release|Any CPU
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.AppStore|Any CPU.Build.0 = Release|Any CPU
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.AppStore|ARM.ActiveCfg = Release|ARM
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.AppStore|ARM.Build.0 = Release|ARM
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.AppStore|iPhone.ActiveCfg = Release|Any CPU
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.AppStore|iPhone.Build.0 = Release|Any CPU
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.AppStore|iPhoneSimulator.ActiveCfg = Release|Any CPU
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.AppStore|iPhoneSimulator.Build.0 = Release|Any CPU
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.AppStore|Mixed Platforms.ActiveCfg = Release|x86
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.AppStore|Mixed Platforms.Build.0 = Release|x86
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.AppStore|x64.ActiveCfg = Release|x64
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.AppStore|x64.Build.0 = Release|x64
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.AppStore|x86.ActiveCfg = Release|x86
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.AppStore|x86.Build.0 = Release|x86
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.Debug|ARM.ActiveCfg = Debug|ARM
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.Debug|ARM.Build.0 = Debug|ARM
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.Debug|iPhone.ActiveCfg = Debug|Any CPU
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.Debug|iPhone.Build.0 = Debug|Any CPU
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.Debug|iPhoneSimulator.ActiveCfg = Debug|Any CPU
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.Debug|iPhoneSimulator.Build.0 = Debug|Any CPU
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.Debug|Mixed Platforms.ActiveCfg = Debug|x86
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.Debug|Mixed Platforms.Build.0 = Debug|x86
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.Debug|x64.ActiveCfg = Debug|x64
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.Debug|x64.Build.0 = Debug|x64
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.Debug|x86.ActiveCfg = Debug|x86
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.Debug|x86.Build.0 = Debug|x86
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.Release|ARM.ActiveCfg = Release|ARM
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.Release|ARM.Build.0 = Release|ARM
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.Release|iPhone.ActiveCfg = Release|Any CPU
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.Release|iPhone.Build.0 = Release|Any CPU
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.Release|iPhoneSimulator.ActiveCfg = Release|Any CPU
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.Release|iPhoneSimulator.Build.0 = Release|Any CPU
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.Release|Mixed Platforms.ActiveCfg = Release|x86
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.Release|Mixed Platforms.Build.0 = Release|x86
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.Release|x64.ActiveCfg = Release|x64
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.Release|x64.Build.0 = Release|x64
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.Release|x86.ActiveCfg = Release|x86
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C}.Release|x86.Build.0 = Release|x86
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -720,6 +778,7 @@ Global
 		{505CEE6A-E069-44BE-9DD2-4F609C1A2FC4} = {AB1FA939-5161-4DDF-8470-A44658E86BC2}
 		{A28B6D00-5285-4F84-A043-8ABAB789A83F} = {AB1FA939-5161-4DDF-8470-A44658E86BC2}
 		{F148873D-B85E-41C3-BE48-C69134F5F1CF} = {9705BAF6-DD11-40F6-B338-E6929B948FF2}
+		{35E6DE23-923D-40CE-B0CD-9CA27638FA5C} = {AB1FA939-5161-4DDF-8470-A44658E86BC2}
 	EndGlobalSection
 	GlobalSection(MonoDevelopProperties) = preSolution
 		version = 0.8.0.0
